% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareTools.R
\name{bblogist.sim}
\alias{bblogist.sim}
\title{Simulates data picking from a beta binomial}
\usage{
bblogist.sim(mles, days, Dmat)
}
\arguments{
\item{Dmat}{}
}
\value{

}
\description{
Simulates data picking from a beta binomial
}
