% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexss.bootstrappedMLEs.R
\name{flexss.bootstrappedMLEs}
\alias{flexss.bootstrappedMLEs}
\title{Maximum Likelihood estimation of the SS model parameters along with Parametric
Bootstrap confidence intervals.}
\usage{
flexss.bootstrappedMLEs(
  problem.counts,
  nrepsi,
  Dmat,
  gens,
  B = 1000,
  neg.cost = TRUE
)
}
\arguments{
\item{problem.counts}{is a matrix.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies.  One column per replicated platings.}

\item{nrepis}{The number of replicated platings. Must be equal to ncol(day.counts)-1}

\item{gen}{Number of generations per day}

\item{Dmat:}{A matrix whose number of rows is the number of days that the assay was ran.  The number of columns corresponds to the number of replicated platings. The entries correspond to the total number of colonies picked for testing in a particular day and replicate.}
}
\value{
A list with the BIC, the negative log-likelihood and a matrix with the model MLEs along with 95\% Parametric boostrap confidence intervals
}
\description{
Maximum Likelihood estimation of the SS model parameters along with Parametric
Bootstrap confidence intervals.
}
\examples{
flexss.bootstrappedMLEs(problem.counts, nrepsi, Dmat, gens, B=1000, neg.cost=TRUE)
}
