% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT.predict.R
\name{HT.predict}
\alias{HT.predict}
\title{Computing the fraction over time of the plasmid-free cells in a stability assay}
\usage{
HT.predict(beta.o, lam, sig, theta, gam, days, gen)
}
\arguments{
\item{beta.o}{Initial fraction of plasmid free cells}

\item{lam}{Segregation rate (a number between 0 and 1)}

\item{sig}{plasmid cost. Usually sig>0 but if -1<sig<0, then this parameter represents the benefit of having a plasmid}

\item{theta}{fraction of plasmid free cells at which the horizontal transfer attains half its maximum}

\item{gam}{Horizontal transfer probability}

\item{days}{vector}

\item{gen}{Number of generations per day}
}
\description{
Function to compute the fraction over time of the plasmid-free cells in a stability assay
assuming that Horizontal transfer is taking place
See Ponciano et al 2007.
}
