% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.negllSS.flex.R
\name{joint.negllSS.flex}
\alias{joint.negllSS.flex}
\title{joint likelihod from a competition experiment and initial fraction of 0.5 and from a stability assay}
\usage{
joint.negllSS.flex(
  parm,
  data.compet,
  data.stabexp,
  nrc,
  nrepexp,
  gen,
  neg.cost = TRUE
)
}
\arguments{
\item{data.compet}{is a matrix for the competition experiment.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies (one column per replicated platings), followed by the same number of columns with the total number of colonies picked per replicate, per day.}

\item{data.stabexp}{is a matrix for the stability experiment.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies (one column per replicated platings), followed by the same number of columns with the total number of colonies picked per replicate, per day.}

\item{nrc}{The number of replicated platings for the competition experiment.}

\item{nrepexp}{The number of replicated platings for the stability experiment.}

\item{gen}{Number of generations per day}

\item{neg.cost}{Logical. If TRUE, then assumes cost is between -1 and 0.}

\item{parm=c(beta.o, lam, sig)}{which are the initial fraction of plasmid free cells, segregation rate and the plasmid cost. All three parameters must be specified according to the inverse transform of either the 1/(1+e^(-x)) or the -1/(1+e^(-x)) functions.}
}
\value{
the negative log-likelihood
}
\description{
Uses data from a competition experiment starting with an initial fraction of 0.5 and from a stability assay
to compute the joint negative log-likelihood for the Segregation Selection model, under binomial sampling
}
\examples{
joint.negllSS.flex(parm, data.compet, data.stabexp,nrc,nrepexp, gen,neg.cost=TRUE)
}
