% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareTools.R
\name{joint.vs.sep.fit}
\alias{joint.vs.sep.fit}
\title{Comparisson between fits}
\usage{
joint.vs.sep.fit(
  datamatAnc,
  datamatEv,
  nrepsAnc = 3,
  nrepsEv = 3,
  my.guess = c(0.5, 0.6, 0.05, log(1.5)),
  my.method = "BFGS",
  plot.Anc = FALSE,
  plot.Ev = FALSE
)
}
\arguments{
\item{plot.Ev}{}
}
\value{

}
\description{
Comparisson between fits
}
