% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointflexssMLEs.R
\name{jointflexssMLEs}
\alias{jointflexssMLEs}
\title{Joint Maximum Likelihood estimation of the SS model parameters along with Parametric
Bootstrap confidence intervals for competition and stability assay data.}
\usage{
jointflexssMLEs(
  data.compet,
  data.stabexp,
  nrc,
  nrepexp,
  gen,
  B = 1000,
  neg.cost = TRUE
)
}
\arguments{
\item{data.compet}{is a matrix for the competition experiment.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies (one column per replicated platings), followed by the same number of columns with the total number of colonies picked per replicate, per day.}

\item{data.stabexp}{is a matrix for the stability experiment.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies (one column per replicated platings), followed by the same number of columns with the total number of colonies picked per replicate, per day.}

\item{nrc}{The number of replicated platings for the competition experiment.}

\item{nrepexp}{The number of replicated platings for the stability experiment.}

\item{gen}{Number of generations per day}

\item{problem.counts}{is a matrix.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies.  One column per replicated platings.}
}
\value{
A list with the BIC, the negative log-likelihood and a matrix with the model MLEs along with 95\% Parametric boostrap confidence intervals
}
\description{
Joint Maximum Likelihood estimation of the SS model parameters along with Parametric
Bootstrap confidence intervals for competition and stability assay data.
}
\examples{
jointflexssMLEs(data.compet=compet.data, data.stabexp= exp.data, nrc=3,nrepexp=3, gen=10, B=1000, neg.cost=TRUE)
}
