% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.time.R
\name{loss.time}
\alias{loss.time}
\title{Function to compute the time (in generations) it takes for a plasmid to be lost
or quasi-extinct.}
\usage{
loss.time(beta.o, lam, sig, crit.pct, stop.when = 300)
}
\arguments{
\item{beta.o}{is the initial fraction of plasmid-free cells}

\item{lam}{is the segregation rate}

\item{sig}{is the cost}

\item{crit.pct}{is the quasi-extinction threshold, typically very close to 1}

\item{stop.when}{is a safe-guard stopping limit:  if the critical threshold has not been reached by 'stop.when' generations, the function stops accumulating generations}
}
\description{
Function to compute the time (in generations) it takes for a plasmid to be lost
or quasi-extinct.
}
