% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareTools.R
\name{negll.bblogist}
\alias{negll.bblogist}
\title{Log likelihood for beta binomial model}
\usage{
negll.bblogist(guess, datamat, nreps = 3)
}
\arguments{
\item{nreps}{}
}
\value{

}
\description{
Log likelihood for beta binomial model
}
\examples{

}
