% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negllSS.flex.R
\name{negllSS.flex}
\alias{negllSS.flex}
\title{Computes the negative log-likelihood for the Segregation Selection model}
\usage{
negllSS.flex(parm, day.counts, nreps, gen, D1, neg.cost = TRUE)
}
\arguments{
\item{day.counts}{is a matrix.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies.  One column per replicated platings.}

\item{nreps}{The number of replicated platings. Must be equal to ncol(day.counts)-1}

\item{gen}{Number of generations per day}

\item{neg.cost}{Logical. If TRUE, then assumes cost is between -1 and 0.}

\item{parm=c(beta.o, lam, sig)}{which are the initial fraction of plasmid free cells, segregation rate and the plasmid cost. All three parameters must be specified according to the inverse transform of either the 1/(1+e^(-x)) or the -1/(1+e^(-x)) functions.}

\item{D1:}{Either a matrix or a vector.  If a matrix, its number of rows is the number of days that the assay was ran.  The number of columns corresponds to the number of replicated platings. The entries correspond to the total number of colonies picked for testing in a particular day and replicate. If a vector, it is assumed that the daily number of colonies picked for each replicate is the same.}
}
\value{
the negative log-likelihood
}
\description{
Computes the negative log-likelihood for the Segregation Selection model, under binomial sampling
Differs from the negloglikeSS.R function in that it allows for a negative cost (a benefit)
hence the 'flex' (flexible) in the name...
See De Gelder et al 2004, 2008 and Ponciano et al 2007.
}
\examples{
negllSS.flex(parm, day.counts, nreps, gen,D1, neg.cost=TRUE)
}
