% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglikeHT.comb.R
\name{negloglikeHT.comb}
\alias{negloglikeHT.comb}
\title{Joint likelihood for two different stability assays are the same, assuming Horizontal Transfer}
\usage{
negloglikeHT.comb(parm, day.counts1, day.counts2, nreps1, nreps2, gen, D1, D2)
}
\description{
Joint likelihood assuming that the parameters for two different stability assays are the same
and assuming that horizontal transfer occurs
For a description of the arguments see the 'negloglikeHT.R' documentation
Mostly used internally by the function 'parwise.fitHT.R'
}
