% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglikeHT.R
\name{negloglikeHT}
\alias{negloglikeHT}
\title{Computes the negative log-likelihood for the Horizontal Transfer model}
\usage{
negloglikeHT(parm, day.counts, nreps, gen, D1)
}
\arguments{
\item{day.counts}{is a matrix.  The first column are the days, the next columns are the counts over time of the plasmid-free colonies.  One column per replicated platings.}

\item{nreps}{The number of replicated platings. Must be equal to ncol(day.counts)-1}

\item{gen}{Number of generations per day}

\item{parm=c(beta.o, lam, sig, }{theta, gamma) which are the initial fraction of plasmid free cells, segregation rate and the plasmid cost. All three parameters must be specified according to the inverse transform of either the 1/(1+e-x) or the exp(x) functions. See example below}

\item{D1:}{Either a matrix or a vector.  If a matrix, its number of rows is the number of days that the assay was ran.  The number of columns corresponds to the number of replicated platings. The entries correspond to the total number of colonies picked for testing in a particular day and replicate. If a vector, it is assumed that the daily number of colonies picked for each replicate is the same.}
}
\value{
the negative log-likelihood
}
\description{
Computes the negative log-likelihood for the Horizontal Transfer model, under binomial sampling
See Ponciano et al 2007.
}
