% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglikeSS.comb.R
\name{negloglikeSS.comb}
\alias{negloglikeSS.comb}
\title{Joint likelihood for two different stability assays}
\usage{
negloglikeSS.comb(parm, day.counts1, day.counts2, nreps1, nreps2, gen, D1, D2)
}
\description{
Joint likelihood assuming that the parameters for two different stability assays are the same
For a description of the arguments see the 'negloglikeSS.R' documentation
Mostly used internally by the function 'parwise.fitSS.R'
}
