% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.predict.R
\name{SS.predict}
\alias{SS.predict}
\title{Computes the solution of the difference equation for the fraction
of plasmid-free cells for the Segregation Selection model.}
\usage{
SS.predict(beta.o, lam, sig, days, gen)
}
\arguments{
\item{beta.o}{Initial fraction of plasmid free cells}

\item{lam}{Segregation rate (a number between 0 and 1)}

\item{sig}{plasmid cost. Usually sig>0 but if -1<sig<0, then this parameter represents the benefit of having a plasmid}

\item{days}{vector}

\item{gen}{Number of generations per day}
}
\value{
a vector of same length as 'says' with the fraction of plasmid free cells
}
\description{
Computes the solution of the difference equation for the fraction
of plasmid-free cells for the Segregation Selection model.
The output is a time series of the fraction of plasmid-free cells
See De Gelder et al 2004, 2008 and Ponciano et al 2007.
}
\examples{
SS.predict(beta.o=0.01, lam=0.003,sig=0.4, days=7, gen=10)
}
