% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSdatasim.R
\name{vs.sim}
\alias{vs.sim}
\title{Data simulation under the variable selection model}
\usage{
vs.sim(
  lam = 0.001533,
  mu = 0.011666,
  tausq = 0.158333,
  days,
  x0 = 0.0066,
  gen = 10,
  Dmat = matrix(50, ncol = nreps, nrow = ndays)
)
}
\arguments{
\item{lam}{lambda parameter in the model}

\item{mu}{mu parameter in the model}

\item{tausq}{tau square parameter}

\item{days}{number of days}

\item{x0}{day zero}

\item{gen}{number of generations per day}

\item{Dmat}{starting point}
}
\value{

}
\description{
Data simulation under the variable selection model
}
\examples{

ndays <-22
days.vec <- 0:(ndays-1)
nreps <- 3
D <- matrix(50, ncol=nreps, nrow=ndays)
varselsim <- vs.sim(lam=0.001533, mu=0.011666, tausq=0.158333, days=days.vec, x0 = 0.0066, gen=10, Dmat=D)


}
