% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSdatasim.R
\name{vs.sim}
\alias{vs.sim}
\title{Data simulation under the variable selection model}
\usage{
vs.sim(
  lam = 0.001533,
  mu = 0.011666,
  tausq = 0.158333,
  days,
  x0 = 0.0066,
  gen = 10,
  Dmat = matrix(50, ncol = nreps, nrow = ndays)
)
}
\arguments{
\item{Dmat}{}
}
\value{

}
\description{
Data simulation under the variable selection model
}
\examples{


}
